requirements.txt
Pandas 
plotly_express
streamlit